+++ src/views/browse/Browse.js
@@ -1,20 +1,6 @@
-import React, { useEffect } from 'react';
-import { useDispatch } from 'react-redux';
-import { getAllClients, getAllDevs, getAllTeams } from '../../actions';
+import React from 'react';
 
 const Browse = props => {
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    if(localStorage.getItem('role') === 'dev') {
-      dispatch(getAllDevs());
-      dispatch(getAllTeams());
-    } else {
-      dispatch(getAllClients());
-      dispatch(getAllTeams());
-    }
-  }, []);
-
   return(
     <div>
       <button onClick={() => props.history.push('/browse/teams')}>Teams</button>
